// generated/avatars.d.ts
// GENERATED DO NOT EDIT
// This source data is located at '..\src\opcodes\avatars.jsonc'
// run `npm run filetypes` to rebuild

export type avatars = {
	gender: number,
	player: {
		slots: [
			number,
			number,
			number,
			number,
			number,
			number,
			number,
			number,
			number,
			number,
			number,
			number,
			number,
			number,
			number,
			number,
		],
		rest: Uint8Array,
	} | null,
	npc: {
		id: number,
		buf: Uint8Array,
		unkff: number,
	} | null,
};